Update basic.tsx
This line breaks upon date selection when value is set in the props, for example: value={Date.parse(date)} with date being a valid date string. When I add a check to see if the getTime function exists, the component seems to work correctly. Maybe this works for some other type of value prop that I'm unaware of?